refactor: use create writing custom hook
